+++ main/src/main/java/com/ldg/main/Models/Role.java
@@ -7,21 +7,21 @@ import javax.persistence.*;
 public class Role {
     public Role(){
     }
-    public Role(Integer id) {
+    public Role(long id) {
         this.id = id;
     }
     @Id
     @GeneratedValue(strategy = GenerationType.AUTO)
     @Column(name = "id")
-    private Integer id;
+    private long id;
 
     @Column(name = "title")
     private String title;
     
-    public Integer getId() {
+    public long getId() {
         return id;
     }
-    public void setId(Integer id) {
+    public void setId(long id) {
         this.id = id;
     }
     public String getTitle() {
